refactor: move options types under types folder
